Add automatic download of the ROM file from the GitHub.
This commit introduces a Makefile in the data directory for downloading
the Gigatron TTL official rom and converting it to the hex format, so
that it can be used by readmemh() function in Verilog.

In addition, updated the README.md file with some of the pre-requisites
for compiling this Verilog model.